added confirm delete with swal

+++ server/public/client.js
@@ -7,9 +7,8 @@ function onReady() {
 
 function addClickListeners() {
     $('#submit-task-btn').on('click', addTask);
-    $('#task-display').on('click', '.mark-done-btn', toggleDone);
-    $('#task-display').on('click', '.delete-task-btn', deleteTask);
-    $('#task-complete-display').on('click', '.mark-done-btn', toggleDone)
+    $('#all-content').on('click', '.mark-done-btn', toggleDone);
+    $('#all-content').on('click', '.delete-task-btn', deleteTask);
 }
 
 function addTask() {
@@ -64,14 +63,23 @@ function toggleDone() {
 
 function deleteTask() {
     const id = $(this).data('id');
-
-    $.ajax({
-        type: 'DELETE',
-        url: `/tasks/${id}`
-    }).then(() => {
-        getTasks();
-    }).catch((err) => {
-        console.log('could not delete', err)
+    Swal.fire({
+        title: `Are you sure you want to delete ${id}?`,
+        showCancelButton: true,
+        confirmButtonText: 'Delete',
+    }).then((result) => {
+        if (result.isConfirmed) {
+            $.ajax({
+                type: 'DELETE',
+                url: `/tasks/${id}`
+            }).then(() => {
+                getTasks();
+            }).catch((err) => {
+                console.log('could not delete', err)
+            })
+        } else if (result.isDenied) {
+            console.log('delete canceled');
+        }
     })
 }
